requirements.txt
alembic==1.10.2
attrs==22.2.0
Babel==2.12.1
bcrypt==4.0.1
blinker==1.5
boltons==21.0.0
bracex==2.3.post1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-option-group==0.5.5
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.3.0
email-validator==1.3.1
face==22.0.0
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
Flask-WTF==1.1.1
glom==22.1.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
peewee==3.16.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.5
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
pytz==2022.7.1
requests==2.28.2
rich==13.3.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semgrep==1.14.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.5.post1
SQLAlchemy-Continuum==1.3.14
SQLAlchemy-Utils==0.40.0
tomli==2.0.1
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
wcmatch==8.4.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
